JWT authentication servies for Company and Employee finished
Two microservices has been implemented: one to manage company
authentication and one to manage employee authentication.

A module named "jwt" has been created and allow other module
to use the JWT authentication.

A module "model" has been created which contains all data
model classes.

The token generation is done via a RestController which provide
a "/token" entrypoint. On this entrypoint, the authentication
must be done with a Basic Http authentication. The result is a
String which is the JWT token generated and which must be used
for other entrypoint.

To refresh a token, a RestController provide a "/token/refresh"
entrypoint. On this entrypoint, the authentication must be done
with a Bearer Http authentication (which contains a JWT token).
The result is a String which is a new JWT token.

Add Employee and Company authentication UT.