Employee authentication UI finished
Employee can now authenticate themselves via the UI.
They must provide their company name, their employee
email and their password.

After that, the employee is redirected to the application.
The application consider the employee connected while the
jwt token that it received when he has been authenticated
is still valid and not expired.

The jwt token is store in the local storage with the key
"jwt_token". Many component at their initialization verify
if the token is always correct and if the employee can always
be considered has authenticated. If it not the case, the
application redirect the employee to the login page.

Langage translation has been added, the component which allow
to change the langage is always display no matter the page.
The component is here just for test and is form and emplacement
will change during the development.