Create the main Application page
After that the employee has been connected, he is
redirected to the application page which allow him
to use Celestial Diary.

For the moment, only the side bar has been
implemented and allow the logout and show
some account information.

The side bar can be wrapped and in this case
only displays icons.